refactor: [#939] change config version
The current config version is 2 leaving the previous before the config
averhaul as version 1.
