+++ packages/configuration/src/v2/network.rs
