+++ packages/configuration/src/v2/tracker_api.rs
@@ -65,7 +65,7 @@ impl HttpApi {
 
 #[cfg(test)]
 mod tests {
-    use crate::v1::tracker_api::HttpApi;
+    use crate::v2::tracker_api::HttpApi;
 
     #[test]
     fn http_api_configuration_should_check_if_it_contains_a_token() {
